Fixed shallow copy in update_groups and updated version.
