+++ modules/platform.py
@@ -1,37 +1,18 @@
 import concurrent.futures
 import copy
 import json
-import logging
-import sys
 import time
 import warnings
 from pathlib import Path
 
-import backoff
 import requests
 
 from vat.vectra import (
-    HTTPException,
     VectraClientV2_5,
     _generate_params,
 )
 
 warnings.filterwarnings("always", ".*", PendingDeprecationWarning)
-
-
-class CustomException(Exception):
-    "Custom Exception raised while failure occurs."
-    pass
-
-
-class TooManyRequestException(Exception):
-    "Custom Exception raised while requests exceeds."
-    pass
-
-
-def kill_process_and_exit(e):
-    logging.error("Exiting current process.")
-    sys.exit()
 
 
 def validate_lte_api_v3_4(func):
@@ -73,152 +54,6 @@ class VectraPlatformClientV3(VectraClientV2_5):
             token=token,
             verify=verify,
         )
-        self.token_headers = {
-            "Content-Type": "application/x-www-form-urlencoded",
-        }
-        self._access = False
-        self._check_token()
-        self.verify = verify
-        self.headers = {
-            "Authorization": f"Bearer {self._access}",
-            "Content-Type": "application/json",
-        }
-
-    def _sleep(self, timeout):
-        time.sleep(timeout)
-
-    @backoff.on_exception(
-        backoff.expo,
-        (
-            requests.exceptions.RequestException,
-            TooManyRequestException,
-            requests.exceptions.HTTPError,
-            CustomException,
-        ),
-        max_tries=5,
-        on_giveup=kill_process_and_exit,
-        max_time=60,
-    )
-    def _refresh_token(self):
-        """Generate access token for API authentication.
-        Returns:
-            str: Access Token
-        """
-        resp = {}
-        logging.info("Generating access token using refresh token.")
-        try:
-            resp = requests.post(
-                url=f"{self.base_url}/oauth2/token",
-                data={
-                    "grant_type": "refresh_token",
-                    "refresh_token": f"{self._refresh}",
-                },
-                headers=self.token_headers,
-                verify=self.verify,
-            )
-            if resp.status_code == 401:
-                raise CustomException("Retrying to generate access token")
-            if resp.status_code == 429:
-                raise TooManyRequestException("Too many requests.")
-            resp.raise_for_status()
-            logging.info("Access token is generated using refresh token.")
-            self._access = resp.json().get("access_token")
-            self._accessTime = int(time.time()) + resp.json().get("expires_in") - 100
-        except CustomException as e:
-            logging.error(f"Error occurred: {e}")
-            self._get_token()
-            raise CustomException
-        except TooManyRequestException as e:
-            logging.info(
-                f"{e}. Retrying after {int(resp.headers.get('Retry-After'))} seconds."
-            )
-            time.sleep(int(resp.headers.get("Retry-After")))
-            raise TooManyRequestException from e
-        except requests.exceptions.HTTPError:
-            logging.error("Vectra API server is down. Retrying after 10 seconds.")
-            time.sleep(10)
-            raise requests.exceptions.HTTPError
-        except requests.exceptions.RequestException as req_exception:
-            logging.error(f"Retrying. An exception occurred: {req_exception}")
-            raise requests.exceptions.RequestException from req_exception
-        except Exception as e:
-            logging.error(f"An exception occurred: {e}")
-
-    @backoff.on_exception(
-        backoff.expo,
-        (
-            requests.exceptions.RequestException,
-            TooManyRequestException,
-            requests.exceptions.HTTPError,
-            CustomException,
-        ),
-        max_tries=5,
-        on_giveup=kill_process_and_exit,
-        max_time=60,
-    )
-    def _get_token(self):
-        """Generate access token for API authentication.
-
-        Returns:
-            str: Access Token
-        """
-        resp = {}
-
-        logging.info("Generating access token.")
-        try:
-
-            resp = requests.post(
-                f"{self.base_url}/oauth2/token",
-                auth=self.auth,
-                headers=self.token_headers,
-                data={"grant_type": "client_credentials"},
-                verify=self.verify,
-            )
-            if resp.status_code == 401:
-                raise CustomException(
-                    f"Status-code {resp.status_code} Exception: Client ID or Client Secret is incorrect."
-                )
-            if resp.status_code == 405:
-                redirect = resp.request.url
-                self.base_url = "https://" + redirect.strip().split("/")[2]
-                self.url = f"{self.base_url}/api/v{self.VERSION3}"
-                self._get_token()
-            if resp.status_code == 429:
-                raise TooManyRequestException("Too many requests.")
-            if resp.status_code == 200:
-                logging.info("Access token is generated.")
-                self._access = resp.json().get("access_token")
-                self._refresh = resp.json().get("refresh_token")
-                self._accessTime = (
-                    int(time.time()) + resp.json().get("expires_in") - 100
-                )
-                self._refreshTime = (
-                    int(time.time()) + resp.json().get("refresh_expires_in") - 100
-                )
-        except CustomException as e:
-            logging.error(f"Error occurred: {e}")
-        except TooManyRequestException as e:
-            logging.info(
-                f"{e}. Retrying after {int(resp.headers.get('Retry-After'))} seconds."
-            )
-            time.sleep(int(resp.headers.get("Retry-After")))
-            raise TooManyRequestException from e
-        except requests.exceptions.HTTPError as e:
-            print(e)
-            logging.error("Vectra API server is down. Retrying after 10 seconds.")
-            time.sleep(10)
-            raise requests.exceptions.HTTPError
-        except requests.exceptions.RequestException as req_exception:
-            logging.error(f"Retrying. An exception occurred: {req_exception}")
-            raise requests.exceptions.RequestException from req_exception
-        except Exception as e:
-            logging.error(f"An exception occurred: {e}")
-
-    def _check_token(self):
-        if not self._access:
-            self._get_token()
-        elif self._accessTime < int(time.time()):
-            self._refresh_token()
 
     def yield_event_results(self, resp, method, **kwargs):
         params = kwargs.get("params", {})
@@ -1634,7 +1469,7 @@ class VectraPlatformClientV3_4(VectraPlatformClientV3_3):
         elif regex := kwargs.get("regex"):
             pass
         else:
-            members = []
+            members = copy.deepcopy(group.get("members", []))
             regex = None
 
         name = kwargs.get("name", group["name"])
